chore(types): adjust types imports
