+++ src/styles/helpers.ts
@@ -1,14 +1,17 @@
-import { Grid, Theme, ThemeProps } from '../types';
+import { Types } from '../types';
 
-function isTheme(props: ThemeProps): props is Theme {
-  return (props as { theme: Theme }).theme === undefined;
+function isTheme(props: Types.ThemeProps): props is Types.Theme {
+  return (props as { theme: Types.Theme }).theme === undefined;
 }
 
-function getTheme(props: ThemeProps): Theme {
+function getTheme(props: Types.ThemeProps): Types.Theme {
   return isTheme(props) ? props : props.theme;
 }
 
-export function mergeThemes(standard: Grid, custom: ThemeProps): Grid {
+export function mergeThemes(
+  standard: Types.Grid,
+  custom: Types.ThemeProps
+): Types.Grid {
   const theme = getTheme(custom);
 
   return {
